First Game Action -> Deal some cards
Deals cards to players using an immutable (but very lightweight) game
state. This actually completely disregards most of the leadup work about
models and constructors and immutability.

To get things moving, I want to just focus on working code and trust
that unit tests and other conventions will save me from problems with
immutability. If I do need that in the future, there will come an
obvious time to tackle that. Without language support, it is taking on
too much too soon.