Remove license going forward
Just because this is public right now does not mean that I necessarily
want anybody using it. As per choosealicense.com, I want to go with no
licensing options for the time being, so that in the future when I do
want to figure it out, I can select the right one without painting
myself into a corner.